refactor(runtime): move identity helpers back into runtime package
