+++ ddtrace/internal/core/crashtracking.py
@@ -11,10 +11,10 @@ from ddtrace import version
 from ddtrace.internal import excepthook
 from ddtrace.internal import forksafe
 from ddtrace.internal import process_tags
-from ddtrace.internal._identity import get_runtime_id
-from ddtrace.internal._identity import on_runtime_id_change
 from ddtrace.internal.compat import ensure_text
 from ddtrace.internal.logger import get_logger
+from ddtrace.internal.runtime import get_runtime_id
+from ddtrace.internal.runtime import on_runtime_id_change
 from ddtrace.internal.settings import env
 from ddtrace.internal.settings._agent import config as agent_config
 from ddtrace.internal.settings.crashtracker import config as crashtracker_config
