+++ ddtrace/internal/runtime/runtime_metrics.py
@@ -3,8 +3,8 @@ from typing import ClassVar  # noqa:F401
 from typing import Optional  # noqa:F401
 
 from ddtrace.internal import atexit
-from ddtrace.internal._identity import on_runtime_id_change
 from ddtrace.internal.constants import EXPERIMENTAL_FEATURES
+from ddtrace.internal.runtime import on_runtime_id_change
 from ddtrace.internal.settings._agent import config as agent_config
 from ddtrace.internal.settings._config import config
 from ddtrace.internal.threads import Lock
