+++ ddtrace/internal/symbol_db/remoteconfig.py
@@ -1,7 +1,6 @@
 import os
 import typing as t
 
-from ddtrace.internal._identity import get_ancestor_runtime_id
 from ddtrace.internal.forksafe import get_generation
 from ddtrace.internal.forksafe import has_forked
 from ddtrace.internal.ipc import SharedStringFile
@@ -10,6 +9,7 @@ from ddtrace.internal.native import RemoteConfigProduct
 from ddtrace.internal.remoteconfig import Payload
 from ddtrace.internal.remoteconfig import RCCallback
 from ddtrace.internal.remoteconfig.worker import remoteconfig_poller
+from ddtrace.internal.runtime import get_ancestor_runtime_id
 from ddtrace.internal.symbol_db.symbols import SymbolDatabaseUploader
